refactor: presenterTest 패키지 구분
