+++ app/src/test/java/woowacourse/shopping/presentation/productlist/ProductListPresenterTest.kt
@@ -1,4 +1,4 @@
-package woowacourse.shopping
+package woowacourse.shopping.presentation.productlist
 
 import io.mockk.every
 import io.mockk.mockk
@@ -6,9 +6,10 @@ import io.mockk.slot
 import io.mockk.verify
 import org.junit.Before
 import org.junit.Test
+import woowacourse.shopping.CartProductInfo
+import woowacourse.shopping.Price
+import woowacourse.shopping.Product
 import woowacourse.shopping.presentation.mapper.toPresentation
-import woowacourse.shopping.presentation.productlist.ProductListContract
-import woowacourse.shopping.presentation.productlist.ProductListPresenter
 import woowacourse.shopping.repository.CartRepository
 import woowacourse.shopping.repository.ProductRepository
 import woowacourse.shopping.repository.RecentProductRepository
